backend/requirements.txt
django
psycopg2
djangorestframework
markdown
django-filter
python-decouple
pyrebase4
firebase-admin
django-cors-headers
whitenoise
drf-yasg
setuptools
django-environ
gunicorn
python-dotenv
requests
pylint
unpack